requirements.txt
torch==2.2.1
torchvision==0.17.1
opencv-python==4.9.0.80
flask==3.0.2
numpy==1.26.4
pandas==2.2.1
pillow==10.2.0
scikit-learn==1.4.1.post1